+++ SwiftUIList/SwiftUIList/MainView/View/ShowMainView.swift
@@ -40,23 +40,3 @@ struct ContentView_Previews: PreviewProvider {
     }
 }
 
-struct ShowCell: View {
-    let show: Show
-    var isLast: Bool
-    @ObservedObject var viewModel: ShowViewModel
-    
-    var body: some View {
-        NavigationLink(destination: ShowDetailView(show: show)) {
-            HStack(spacing: 20){
-                ShowMainImage(urlString: show.image.original)
-                Text(show.name)
-                    .font(.title2)
-                
-                if self.isLast == true {
-                    Text("Last")
-                }
-            }
-        }
-    }
-}
-
